requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==19.1.0
beautifulsoup4==4.12.2
boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
cssselect==1.2.0
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-ses==3.5.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
fake-useragent==1.1.3
frozenlist==1.4.0
future==0.18.3
Google-Search-API==1.1.14
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
lxml==4.9.3
markdown2==2.4.9
maxminddb==2.4.0
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
Pillow==10.0.0
playwright==1.36.0
proxy-checker==0.6
proxybroker==0.3.2
psycopg2==2.9.6
pycares==4.3.0
pycparser==2.21
pycurl==7.45.2
pyee==9.0.4
PyJWT==2.7.0
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
readtime==2.0.0
requests==2.30.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
selenium==2.53.6
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.5.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.15
vcrpy==5.0.0
wrapt==1.15.0
yarl==1.9.2
